migrate showcase from vuu-layout to grid-layout
