+++ vuu-ui/packages/grid-layout/src/GridLayout.css
@@ -20,6 +20,12 @@
   &.vuuResizing {
     transition: .3s;
   }
+
+  &.vuuFullPage {
+    height: 100vh;
+    width: 100vw;
+  }
+  
 }
 
 .vuuGridLayoutItem {
