+++ vuu-ui/tools/vuu-showcase/src/showcase-main/ShowcaseShell.css
@@ -38,6 +38,7 @@ body {
 
 .ShowcaseContent {
   background-color: var(--salt-container-primary-background);
+  height: 100%;
   overflow: hidden;
 }
