+++ vuu-ui/tools/vuu-showcase/src/showcase-main/iframe/IFrame.tsx
@@ -1,3 +1,4 @@
+import React from "react";
 import { Density, ThemeMode } from "@finos/vuu-utils";
 
 export interface IFrameProps {
